Migrate theme builder from context to component: - Add component for theme builder - Remove theme builder context due to performance issue - Add theme create page - Add next router mock in test files
