+++ components/editor/SnippngCodeArea.tsx
@@ -141,14 +141,15 @@ const SnippngCodeArea: React.FC<Props> = ({ underConstructionTheme }) => {
 
   useEffect(() => {
     // if there is a uid means we are on edit page where we want to avoid persisting the editor config
-    if (uid) return;
+    // underConstructionTheme means user is on the theme page which we don't want to persist
+    if (uid || underConstructionTheme) return;
     // persist the editor config changes only when user is creating new snippet
     LocalStorage.set("config", {
       ...editorConfig,
       uid: undefined,
       ownerUid: undefined,
     });
-  }, [editorConfig, uid]);
+  }, [editorConfig, uid, underConstructionTheme]);
 
   return (
     <>
@@ -160,7 +161,10 @@ const SnippngCodeArea: React.FC<Props> = ({ underConstructionTheme }) => {
         <NoSSRWrapper>
           <div className="rounded-md bg-white dark:bg-zinc-900 md:p-8 p-4 flex justify-center border-[1px] flex-col items-center dark:border-zinc-500 border-zinc-200 shadow-md w-full">
             <div className="w-full">
-              <SnippngControlHeader wrapperRef={wrapperRef} />
+              <SnippngControlHeader
+                creatingTheme={!!underConstructionTheme}
+                wrapperRef={wrapperRef}
+              />
             </div>
             {bgImageVisiblePatch ? (
               <button
@@ -261,65 +265,63 @@ const SnippngCodeArea: React.FC<Props> = ({ underConstructionTheme }) => {
                 ) : null}
               </div>
             </div>
-            {!underConstructionTheme ? (
-              <div className="w-full mt-8 flex md:flex-row flex-col gap-4 justify-start items-center">
-                <div className="w-full">
-                  <Input
-                    value={snippetsName}
-                    onChange={(e) =>
-                      handleConfigChange("snippetsName")(e.target.value)
-                    }
-                    placeholder="Snippet name..."
-                  />
-                </div>
-                <div className="flex flex-shrink-0 gap-4 md:flex-row flex-col md:w-fit w-full">
+            <div className="w-full mt-8 flex md:flex-row flex-col gap-4 justify-start items-center">
+              <div className="w-full">
+                <Input
+                  value={snippetsName}
+                  onChange={(e) =>
+                    handleConfigChange("snippetsName")(e.target.value)
+                  }
+                  placeholder="Snippet name..."
+                />
+              </div>
+              <div className="flex flex-shrink-0 gap-4 md:flex-row flex-col md:w-fit w-full">
+                <Button
+                  id="save-snippet-btn"
+                  StartIcon={ArrowDownOnSquareStackIcon}
+                  disabled={saving}
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    if (!user)
+                      return addToast({
+                        message: "Please login first",
+                        type: "error",
+                        description:
+                          "You need to login before saving the snippet",
+                      });
+                    if (!snippetsName)
+                      return addToast({
+                        message: "Snippet name is required",
+                        type: "error",
+                      });
+                    else saveSnippet();
+                  }}
+                >
+                  {saving
+                    ? "Saving..."
+                    : uid // if there is a uid, we are on snippet details page where user can copy the snippet
+                    ? "Fork snippet"
+                    : "Save snippet"}
+                </Button>
+                {uid && user && user.uid === ownerUid ? (
                   <Button
-                    id="save-snippet-btn"
-                    StartIcon={ArrowDownOnSquareStackIcon}
-                    disabled={saving}
+                    StartIcon={ArrowPathIcon}
+                    disabled={updating}
                     onClick={(e) => {
                       e.stopPropagation();
-                      if (!user)
-                        return addToast({
-                          message: "Please login first",
-                          type: "error",
-                          description:
-                            "You need to login before saving the snippet",
-                        });
                       if (!snippetsName)
                         return addToast({
                           message: "Snippet name is required",
                           type: "error",
                         });
-                      else saveSnippet();
+                      updateSnippet();
                     }}
                   >
-                    {saving
-                      ? "Saving..."
-                      : uid // if there is a uid, we are on snippet details page where user can copy the snippet
-                      ? "Fork snippet"
-                      : "Save snippet"}
+                    {updating ? "Updating..." : "Update snippet"}
                   </Button>
-                  {uid && user && user.uid === ownerUid ? (
-                    <Button
-                      StartIcon={ArrowPathIcon}
-                      disabled={updating}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        if (!snippetsName)
-                          return addToast({
-                            message: "Snippet name is required",
-                            type: "error",
-                          });
-                        updateSnippet();
-                      }}
-                    >
-                      {updating ? "Updating..." : "Update snippet"}
-                    </Button>
-                  ) : null}
-                </div>
+                ) : null}
               </div>
-            ) : null}
+            </div>
           </div>
           {uid ? (
             <small className="dark:text-zinc-300 text-left text-zinc-600 py-2 inline-block">
